% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enabench.r
\name{enabench}
\alias{enabench}
\title{Fonctional analysis benchmark}
\usage{
enabench(
  symbollist = NULL,
  foldchange = NULL,
  keywords = NULL,
  species = "hs",
  background = 25000,
  overlapmin = 2,
  topena = 50,
  labsize = 11,
  dpibarplot = "screen",
  dolistnetwork = TRUE,
  dogenesetnetwork = FALSE,
  confidence = 0,
  intmaxdh = 6000,
  intmax = 1e+07,
  nodesize = 0.39,
  mings = 5,
  maxgs = 700,
  dotopgo = TRUE,
  ipafile = NULL,
  nodegraph = c(5, 10),
  path = ".",
  dirname = NULL
)
}
\arguments{
\item{symbollist}{character Symbol or NCBI gene ID}

\item{foldchange}{data.frame Symbol list with fold-change values to display}

\item{keywords}{data.frame keywords list for geneset selection}

\item{species}{character hs mm rn dr ss}

\item{background}{numeric}

\item{overlapmin}{numeric for minimum overlap between geneset and list}

\item{topena}{numeric top features to plot}

\item{labsize}{numeric size of function in barplot}

\item{dpibarplot}{character barplot resolution}

\item{dolistnetwork}{logical create symbollist interaction network}

\item{dogenesetnetwork}{logical create top geneset interaction network}

\item{confidence}{numeric MSigDB interaction confidence score}

\item{intmaxdh}{numeric maximum number of interaction to use for Davidson and Harel algorithm layout}

\item{intmax}{numeric maximum number of interaction to use}

\item{nodesize}{numeric change Symbol size}

\item{mings}{numeric minimal size of a gene set}

\item{maxgs}{numeric maximal size of a gene set}

\item{dotopgo}{logical do gene ontology enrichment analysis using topGO}

\item{ipafile}{character Ingenuity Pathways Analysis export all results file}

\item{nodegraph}{numeric number of top GO term to use for GO term nodes graph}

\item{path}{character for relative path of output directory}

\item{dirname}{character name for output}
}
\value{
file with enrichment analysis results
}
\description{
MSigDB enrichment analysis
}
\examples{
# not run
# enabench(symbollist , dotopgo , ipafile)
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
