% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enaf2.r
\name{enaf2}
\alias{enaf2}
\title{MSigDB Enrichment Analysis For Two list.}
\usage{
enaf2(
  list1 = NULL,
  list2 = NULL,
  addlist = NULL,
  listnames = NULL,
  species = "hs",
  top = 50,
  collection = NULL,
  intmaxdh = 5000,
  mings = 5,
  maxgs = 500,
  overlapmin = 2,
  layout = NULL,
  dirname = NULL,
  path = "."
)
}
\arguments{
\item{list1}{character Symbol list 1}

\item{list2}{character Symbol list 2}

\item{addlist}{character additional Symbol list 2}

\item{listnames}{character vector with 3 list names respectively}

\item{species}{character hs mm rn dr ss}

\item{top}{numeric number of top patways to extract}

\item{collection}{character MSigDB collection name to extract top pathways}

\item{intmaxdh}{numeric maximum number of interaction to use for Davidson and Harel algorithm layout}

\item{mings}{numeric minimal size of a gene set}

\item{maxgs}{numeric maximal size of a gene set}

\item{overlapmin}{numeric minimal overlap to keep for gene set analysis}

\item{layout}{numeric for layout neetwork 1 fr by default 2 dh 3 tree 4 circle 5 grid 6 sphere}

\item{dirname}{character output directory name}

\item{path}{character path}
}
\value{
directory
}
\description{
MSigDB Enrichment Analysis For Two list
}
\examples{
# not run
# menaf2( list1 , list2 )
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
