% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipa.r
\name{ipa}
\alias{ipa}
\title{ipa}
\usage{
ipa(
  ipafile = "",
  top = 50,
  labsize = 8,
  dpi = "retina",
  overlapmin = 2,
  enascoremin = 1,
  background = 25000,
  path = ".",
  dirname = NULL
)
}
\arguments{
\item{ipafile}{character path to the file}

\item{top}{numeric top feature to display}

\item{labsize}{numeric feature size}

\item{dpi}{character jpeg resolution between retina print screen}

\item{overlapmin}{numeric for minimum overlap between geneset and list}

\item{enascoremin}{numeric for minimum ratio ena}

\item{background}{numeric}

\item{path}{character for relative path of output directory}

\item{dirname}{character name for output}
}
\value{
no values
}
\description{
parse and create plot from ipa all enrichment results
}
\examples{
# not run
# ipa( file , path "" )
}
\author{
Florent Dumont \href{mailto:florent.dumont@univresite-paris-saclay.fr}{florent.dumont@univresite-paris-saclay.fr}
}
