% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.r
\name{network}
\alias{network}
\title{stringDB interaction network}
\usage{
network(
  nodelist = NULL,
  foldchange = NULL,
  species = NULL,
  confidence = 400,
  intmaxdh = 2500,
  intmax = 5000,
  nodesize = 0.39,
  title = "Network",
  path = ".",
  dirname = NULL
)
}
\arguments{
\item{nodelist}{character list of Symbols to make network}

\item{foldchange}{data.frame Symbol list with fold-change values to display}

\item{species}{character for species hs mm rn dr}

\item{confidence}{numeric for minus interaction confidence score}

\item{intmaxdh}{numeric maximum number of interaction to use for Davidson and Harel algorithm layout}

\item{intmax}{numeric maximum number of interaction to use}

\item{nodesize}{numeric change Symbol size}

\item{title}{character plot title}

\item{path}{character file path}

\item{dirname}{character results directory name}
}
\value{
result folder
}
\description{
stringDB interaction network
}
\examples{
# not run
# network(symbollist)
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
