% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topgo.r
\name{topgo}
\alias{topgo}
\title{topGO analysis}
\usage{
topgo(
  symbollist = NULL,
  species = "hs",
  overlapmin = 1,
  nodegraph = c(5, 10),
  top = 50,
  annot = FALSE,
  path = ".",
  dirname = NULL
)
}
\arguments{
\item{symbollist}{character list of Symbol}

\item{species}{character}

\item{overlapmin}{numeric min overlap between Symbol and GO term}

\item{nodegraph}{numeric number of top GO term to use for GO term nodes graph}

\item{top}{numeric number of GOTerm to display on barplot}

\item{annot}{logical if TRUE (by default) annotate list before analysis}

\item{path}{character file path}

\item{dirname}{character results directory name}
}
\value{
directory with barplot and graph node of GO term
}
\description{
do topgo analysis
}
\details{
topGO analysis using classic fisher method
}
\examples{
# not run
# topgo( Symbollist )
}
\author{
Florent Dumont \href{mailto:florent.dumont@univresite-paris-saclay.fr}{florent.dumont@univresite-paris-saclay.fr}
}
